+++ src/CompilePackage.cs
@@ -1,8 +1,6 @@
-﻿using EnvDTE;
-using EnvDTE80;
-using Microsoft.Build.Construction;
-using Microsoft.VisualStudio.Shell;
+﻿using Microsoft.VisualStudio.Shell;
 using System;
+using System.ComponentModel.Design;
 using System.Runtime.InteropServices;
 using System.Threading;
 using Tasks = System.Threading.Tasks;
@@ -11,18 +9,22 @@ namespace TypeScriptCompileOnSave
 {
     [PackageRegistration(UseManagedResourcesOnly = true)]
     [InstalledProductRegistration("#110", "#112", "1.0", IconResourceID = 400)]
-    [Guid("92479255-4cce-4754-9caa-a0c47a10e055")]
+    [Guid(PackageGuids.guidCompilePackageString)]
+    [ProvideMenuResource("Menus.ctmenu", 1)]
+    [ProvideAutoLoad(PackageGuids.guidAutoLoadString)]
+    [ProvideUIContextRule(PackageGuids.guidAutoLoadString,
+    name: "Auto load",
+    expression: "aspnet & typescript & (js | jsx)",
+    termNames: new[] { "aspnet", "typescript", "js", "jsx" },
+    termValues: new[] { "ActiveProjectCapability:DotNetCoreWeb", "ActiveProjectCapability:TypeScript", "HierSingleSelectionName:.js$", "HierSingleSelectionName:.jsx$" })]
     public sealed class CompilePackage : AsyncPackage
     {
-        //protected override async Tasks.Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
-        //{
-        //    var dte = await GetServiceAsync(typeof(DTE)) as DTE2;
-
-        //    var project = dte.Solution.Projects.Item(1);
-
-        //    var root = ProjectRootElement.Open(project.FullName);
-        //    var proj = new Microsoft.Build.Evaluation.Project(root);
-        //    bool success = proj.Build("CompileTypeScriptWithTSConfig");
-        //}
+        protected override async Tasks.Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
+        {
+            if (await GetServiceAsync(typeof(IMenuCommandService)) is OleMenuCommandService commandService)
+            {
+                AddConfigFile.Initialize(this, commandService);
+            }
+        }
     }
 }