+++ src/Constants.cs
@@ -4,6 +4,7 @@ namespace TypeScriptCompileOnSave
 {
     public class Constants
     {
+        public const string ConfigFileName = "tsconfig.json";
         public static string TscLocation = Environment.ExpandEnvironmentVariables(@"%programfiles(x86)%\Microsoft SDKs\TypeScript\");
 
         public static string[] FileExtensions =
@@ -16,5 +17,14 @@ namespace TypeScriptCompileOnSave
         {
             "{9A19103F-16F7-4668-BE54-9A1E7A4F7556}" // ASP.NET Core
         };
+
+        public const string DefaultTsConfig = @"{{
+  ""include"": [ ""{0}"" ],
+  ""compilerOptions"": {{
+    ""allowJs"": true,
+    ""sourceMap"": true,
+    ""outFile"": ""wwwroot/js/bundle.js""
+  }}
+}}";
     }
 }